// src/types.d.ts
import { TODO_FILTERS } from "./const"

export interface Todo {
    id: string,
    title: string,
    completed: boolean
}
export type TodoId = Pick<Todo,"id">
export type TodoTitle = pick<Todo,"title">
export type TodoCompleted = Pick<Todo,"completed">


export type ListofTodos = Todo[]

export type FilterValues = typeof TODO_FILTERS [keyof typeof TODO_FILTERS]